fix: fix the x00 bug
