--- backend/plugins/jira/tasks/apiv2models/util.go
@@ -1,33 +1,0 @@
-/*
-Licensed to the Apache Software Foundation (ASF) under one or more
-contributor license agreements.  See the NOTICE file distributed with
-this work for additional information regarding copyright ownership.
-The ASF licenses this file to You under the Apache License, Version 2.0
-(the "License"); you may not use this file except in compliance with
-the License.  You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package apiv2models
-
-import (
-	"reflect"
-	"strings"
-)
-
-func stripZeroByte(ifc interface{}) {
-	v := reflect.ValueOf(ifc).Elem()
-	for i := 0; i < v.NumField(); i++ {
-		if v.Field(i).Type() == reflect.TypeOf("") {
-			stripped := strings.ReplaceAll(v.Field(i).String(), "\u0000", "")
-			v.Field(i).Set(reflect.ValueOf(stripped))
-		}
-	}
-}
